downloader_app/requirements.txt
streamlit>=1.28.0
snowflake-snowpark-python>=1.9.0
pandas>=2.0.0
python-dotenv>=1.0.0 